complete lab2 debug and IntList

+++ lab2/DebugPractice/DebugExercise1.java
@@ -4,7 +4,7 @@
  */
 public class DebugExercise1 {
     public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
+        float quotient = (float) top / (float) bottom;
         int result = Math.round(quotient);
         return result;
     }
@@ -23,6 +23,7 @@ public class DebugExercise1 {
         int t3 = 3;
         int b3 = 4;
         int result3 = divideThenRound(t3, b3);
+//        System.out.println(Math.round(0.75));
         System.out.println("round(" + t3 + "/" + b3 + ")=" + result3);
     }
 }
